fix config, move cardinality item impl to relay-metrics
